Refactor PlannedRouteIntegrationTests and added comments
